docs(templates): re-structure for App Shell routing
- merge/simplify templates further
- colocate classes in components
