--- templates/AssetInventory/CardView/index.tsx
@@ -1,1 +1,0 @@
-export * from "./CardView";
