--- templates/AssetInventory/ListView/index.tsx
@@ -1,1 +1,0 @@
-export * from "./ListView";
