--- templates/AssetInventory/styles.tsx
@@ -1,11 +1,0 @@
-import { css } from "@emotion/css";
-import { theme } from "@hitachivantara/uikit-react-core";
-
-const styles = {
-  root: css({
-    display: "grid",
-    gap: theme.space.sm,
-  }),
-};
-
-export default styles;
