--- templates/Dashboard/utils.ts
@@ -1,2 +1,0 @@
-export const formatter = (value?: number | string) =>
-  `${Number(value) / 1000}k`;
