+++ templates/DetailsView/utils.ts
