--- templates/ListView/Indicator/index.ts
@@ -1,1 +1,0 @@
-export * from "./Indicator";
