--- templates/ListView/Kpi/index.ts
@@ -1,1 +1,0 @@
-export * from "./Kpi";
