--- templates/ListView/styles.ts
@@ -1,13 +1,0 @@
-import { css } from "@emotion/css";
-import { theme } from "@hitachivantara/uikit-react-core";
-
-const styles = {
-  paddingTop: css({
-    paddingTop: theme.space.md,
-  }),
-  marginTop: css({
-    marginTop: theme.space.md,
-  }),
-};
-
-export default styles;
