+++ templates/Welcome/index.tsx
@@ -1,15 +1,35 @@
+import { css } from "@emotion/css";
 import {
   HvGrid,
   HvInput,
   HvTypography,
+  theme,
 } from "@hitachivantara/uikit-react-core";
 import { Heart, Palette } from "@hitachivantara/uikit-react-icons";
 
-import styles from "./styles";
+const classes = {
+  root: css({
+    height: "100%",
+    padding: theme.space.xl,
+    background: `linear-gradient(70deg, ${theme.colors.atmo2} 55%, transparent 55%), linear-gradient(180deg, ${theme.colors.brand}, #000)`,
+  }),
+  glossaryContainer: css({
+    margin: theme.spacing("sm", 0),
+    display: "flex",
+    flexDirection: "column",
+    gap: theme.space.xs,
+  }),
+  imageContainer: css({
+    display: "flex",
+    alignItems: "center",
+
+    img: { maxHeight: "450px" },
+  }),
+};
 
 const Welcome = () => {
   return (
-    <div className={styles.root}>
+    <div className={classes.root}>
       <HvGrid container>
         <HvGrid item xs={6}>
           <HvGrid container>
@@ -29,7 +49,7 @@ const Welcome = () => {
               <HvInput type="search" placeholder="What are you looking for?" />
             </HvGrid>
             <HvGrid item xs={6}>
-              <div className={styles.glossaryContainer}>
+              <div className={classes.glossaryContainer}>
                 <Palette iconSize="M" role="none" />
                 <HvTypography variant="title4" component="p">
                   Themeable
@@ -41,7 +61,7 @@ const Welcome = () => {
               </div>
             </HvGrid>
             <HvGrid item xs={6}>
-              <div className={styles.glossaryContainer}>
+              <div className={classes.glossaryContainer}>
                 <Heart iconSize="M" role="none" />
                 <HvTypography variant="title4" component="p">
                   Community
@@ -55,7 +75,7 @@ const Welcome = () => {
           </HvGrid>
         </HvGrid>
         <HvGrid item xs={6} display="flex">
-          <div className={styles.imageContainer}>
+          <div className={classes.imageContainer}>
             <img
               alt="Monitor showing a dashboard built with UI Kit components"
               src="https://i.imgur.com/BPIEkD7.png"
